http2.py: wrap head(), send_and_recv() call in exception-catcher loop
the reason behind this change is that previously only the sending part
of the send/recv transaction was caught, and i experienced a connection
loss during the recv step, which resulted in the termination of the app.

besides, the previous approach to catch only send_raw was bogus:
we must look at both sending and receiving as a complete transaction,
if any of them fails we must start over.
if we would catch send and receive separately, it could happen that
and error happens during receive (e.g. a remote disconnect), so trying
only receiving over and over again would not succeed.

both get() and post() do the entire network-transaction during
send_and_recv(), but head() is special and does things manually, so it
got a new wrapper.

in the future the wrapped function which does not catch the exception
might be made public, so a user who wants to inspect it or do the
handling manually can do so. an alternative approach could be to save
the last exception object in the http component.