add a fetched timestamp to tweets, for retweets timeline order
since retweets only provide the original timestamp, a retweet of
an old tweet will show up way back in the timeline and will likely
not be seen.

as a remedy, we now save a fetched timestamp too.
if existing, and if a retweet, tweets will be sorted first by fetched
timestamp, than by original timestamp.

for the case that we fetch several pages of tweets with a --count
argument > 20, we put the same fetched timestamp on all tweets
downloaded in that invocation in order to prevent older tweets from
showing up as if they were newer.