improve placement of retweets in the timeline
fetching a whole account's tweet history with --count=1, all retweets
went right on top on the timeline, even though they were several yrs
old, due to usage of the 'fetched' timestamp.

(the issue is that twitter delivers only the timestamp of the original
 tweet, but not when it was retweeted).

we now add another field, rid_timestamp which has the timestamp of
the next-newest tweet in the timeline, minus one.

this fixes the issue mostly.