do not show retweets of people in watchlist
if you follow 5 people that know each other, and one of them tweets,
you might end up getting the same tweet 5x in your timeline, if the
other 4 retweet.

the name in the watchlist must have the right spelling (in regard to
upper/lowercase) for it to work.
(the right spelling is the one used in the tweets' title bar).