fix high CPU load while supposedly idling
for some reason, the scrape function reloaded all user's tweets
every single call (i.e. every second), even if no work is to be done.

it's questionable whether the tweets need to be reloaded at all, but
moving the load to before we actually scrape fixes the high CPU load.