Revert "do not show retweets of people in watchlist"
This reverts commit 62e4fd5ca551c7949ee6edb2394989731d3c90ca.

The way this commit is done is bogus: it merely prevents retweets
from *being displayed*, but they're not actually removed from the
lists everything else operates on. so let's say your tpp setting
is 10, and 5 retweets matched the criteria, only 5 tweets would
be displayed on that page.
implemented in this way, this makes computed page numbers
unreliable.