turn unicode into byte strings for md5()
this fixes an exception encountered with this doc
https://www.openssl.org/news/secadv/20170126.txt from
https://twitter.com/robertswiecki/status/824650521436950528
which http2.py "decodes" as utf8 due to content-type charset setting.

since http2 only deals with utf8 it's safe to "encode" all unicode
strings to utf-8 previous to calling md5 on it, which expects a byte
string. everything else is shipped as byte strings already.