make search box remember user setting
strictly speaking, the make_index_link() change is not necessary, because
the form action field does not propagate the GET url parameters contained.
only named fields in the form will end up in the GET request.
i left the change anyways because if at some point we decide to change
"index.html" to something else, we just need the change in a single spot.