always try to fetch until we get to a known tweet in count==-2 mode
it's well possible that we have a network outage lasting for several hours,
and one of our followed users posts more than 20 (1 page) tweets in that time,
or that more than 20 tweets have been made during the scrape timeout.
that means we should always, not just on the first run, try to fetch as many
tweets as necessary until we arrive at a tweet that we already mirrored.