improve sorting of tweets
it's been noticed that apparently twitter uses the timestamp submitted
by the tweeting app of the person making the tweet, instead of the
server's timestamp, which causes very often that the newest tweets
weren't sorted on top of the timeline, but appeared further down
if the tweeter's clock is off.
the only really reliable method to get the sorting right is using
the twat id, but twitter changed its algorithm to produce the id
somewhere between 2016 and 2018.
therefore, we now use the twat id to sort tweets newer than 01/01/2018,
and the old method for everything before that date.