rocksock: fix bug adding target to caller's proxy list
assigning the caller's proxy list to self.proxychain and then adding
the target would actually add it to the caller's proxy list.

this blew up in practice when e.g. http2.py was fetching an URL with
a redirect, and a new rocksock object was created with the existing
proxylist, which now contained a bogus "proxy" (actually the http
host from the previous connection).