remove http server setup code from thread and catch EADDRINUSE
when restarting twatbot from within pdb, *every single time* an exception
due to EADDRINUSE will be raised, even though the server code uses
SO_REUSEADDR (which seems to have no effect).

catching the error in an attempt to loop until the resource is available
is futile, because as long as the pdb process is not terminated, the service
port is not released by the OS and the error will be rethrown ad infinitum.

therefore when the error is detected, we show a message (instead of an ugly
backtrace) and exit with error. this is only possible from the main thread.