fix zlib error -5 when encountering partial zlib data
some m$ IIS crap server, when requesting gzip-encoded data, removes
content-length header, and then first writes a 16 byte chunk with
the gzip header, then additional chunks.

this caused our implementation to read only once with -1 (unknown
size) from the stream, and then only 16 bytes were received.

when we now encounter this situation, we try to decode, and if it
fails with the error -5: incomplete or partial data, we just read
another chunk and try again until it works.

the site/url in question triggering this is:

https://www.infosecurity-magazine.com:443/news/vpns-used-to-defeat-censorship-and-data-retention/?utm_source=twitterfeed&utm_medium=twitter

running: PWS/8.3.1.0.8